Recreated the pins tables.  Added routes for PinHistory to query for the PIN History based on the Pin or get the full collection.

+++ laravel/database/migrations/2020_10_10_224436_create_pins_table.php
@@ -17,8 +17,8 @@ class CreatePinsTable extends Migration
             $table->bigIncrements('id');
             $table->timestamps();
             $table->string('rfidNumber');
-            $table->integer('latitude');
-            $table->integer('longitude');
+            $table->double('latitude', 8, 6);
+            $table->double('longitude', 9, 6);
         });
     }
